Implementing tracing feature - Added possibility to run docker-compose of all tools services and main service and going to http://localhost:16686 Jaeger-UI with checking app traces - Update opentelemetry version - Created feature to print out logs with [Local] or [Dockerized] flag, with fastly understand where project running - Created MDCService for executing and adding mdc.put if need
