+++ app/kafka/src/main/java/com/simple/kafka/handler/SimpleRequestMessageHandlerImpl.java
@@ -9,7 +9,7 @@ import com.simple.kafka.config.SimpleKafkaTopicsProperties;
 import com.simple.models.enums.SimpleKafkaMessageType;
 import com.simple.models.kafka.SimpleKafkaRequest;
 import com.simple.service.api.converter.SimpleKafkaConverter;
-import com.simple.service.api.service.SimpleService;
+import com.simple.service.api.SimpleService;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
