+++ app/service-api/src/main/java/com/simple/service/api/MDCService.java
@@ -1,0 +1,5 @@
+package com.simple.service.api;
+
+public interface MDCService {
+    void putDiagnosticDockerized(final Runnable runnable);
+}
