+++ app/src/main/resources/application.properties
@@ -1,5 +1,6 @@
 spring.application.name=SimpleProcessor
 
+# MongoDB
 spring.data.mongodb.database=tripto
 spring.data.mongodb.username=root
 spring.data.mongodb.password=123456
@@ -8,6 +9,7 @@ spring.data.mongodb.host=${MONGO_HOST:localhost}
 spring.data.mongodb.port=27017
 spring.data.mongodb.auto-index-creation=true
 
+# Prometheus (metrics)
 spring.mvc.throw-exception-if-no-handler-found=true
 management.endpoints.web.exposure.include=prometheus
 management.endpoints.web.cors.allowed-origins=http://localhost:8080
@@ -37,6 +39,5 @@ spring.kafka.producer.bootstrap-servers=${spring.kafka.bootstrap-servers}
 spring.kafka.producer.key-serializer=org.apache.kafka.common.serialization.StringSerializer
 spring.kafka.producer.value-serializer=org.apache.kafka.common.serialization.ByteArraySerializer
 
-# OpenTelemetry
-spring.sleuth.otel.config.trace-id-ratio-based=1.0
-spring.sleuth.otel.exporter.otlp.endpoint=http://collector:4317
+# Self properties
+simple.logging.prefix=${LOGGING_PREFIX:Local}
