+++ gradle.properties
@@ -27,7 +27,7 @@ test_containers_version=1.18.3
 springdoc_openapi_version=2.1.0
 
 # [OpenTelemetry]
-opentelemetry_version=1.19.0
+opentelemetry_version=1.29.0
 
 # [Prometheus - grafana]
 metrics_prometheus_version=1.11.2
